Refactoring: Moved some classes to new packages
