+++ com.apt.project.eshop/src/it/java/com/apt/project/eshop/controller/mongo/EShopControllerIT.java
@@ -1,4 +1,4 @@
-package com.apt.project.eshop.controller;
+package com.apt.project.eshop.controller.mongo;
 
 import static java.util.Arrays.asList;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -18,9 +18,10 @@ import org.mockito.MockitoAnnotations;
 import org.mockito.Spy;
 import org.testcontainers.containers.GenericContainer;
 
+import com.apt.project.eshop.controller.EShopController;
 import com.apt.project.eshop.management.ShopManager;
 import com.apt.project.eshop.management.TransactionManager;
-import com.apt.project.eshop.management.TransactionalShopManager;
+import com.apt.project.eshop.management.mongo.TransactionalShopManager;
 import com.apt.project.eshop.model.Product;
 import com.apt.project.eshop.repository.CartRepository;
 import com.apt.project.eshop.repository.ProductRepository;
