+++ com.apt.project.eshop/src/main/java/com/apt/project/eshop/app/swing/EShopSwingApp.java
@@ -9,7 +9,7 @@ import java.util.logging.Logger;
 
 import com.apt.project.eshop.controller.EShopController;
 import com.apt.project.eshop.management.ShopManager;
-import com.apt.project.eshop.management.TransactionalShopManager;
+import com.apt.project.eshop.management.mongo.TransactionalShopManager;
 import com.apt.project.eshop.model.Product;
 import com.apt.project.eshop.view.swing.EShopSwingView;
 import com.mongodb.MongoClient;
